Refactor PluginsTest
