+++ src/test/java/school/redrover/model/ManageJenkinsPage.java
@@ -1,11 +1,9 @@
 package school.redrover.model;
 
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import school.redrover.model.base.BasePage;
-import school.redrover.model.base.PluginsPage;
 
 public class ManageJenkinsPage extends BasePage {
